feat: 삭제 api에 인증로직 추가

+++ app/(route-handlers)/api/posts/[id]/route.ts
@@ -1,5 +1,7 @@
 import { revalidatePath, revalidateTag } from 'next/cache';
 import { PostsService } from '@/app/_lib/posts/Posts.service';
+import { NextRequest } from 'next/server';
+import { jwtVerify } from 'jose';
 
 export async function GET(
   req: Request,
@@ -19,10 +21,16 @@ export async function GET(
 }
 
 export async function DELETE(
-  req: Request,
+  req: NextRequest,
   { params }: { params: { id: string } },
 ) {
   try {
+    const token = req.cookies.get('token')?.value;
+    if (!token) {
+      throw Error();
+    }
+
+    await jwtVerify(token, new TextEncoder().encode(process.env.JWT_SECRET!));
     const deletedPost = await PostsService.deletePost(params.id);
     revalidatePath('/');
     revalidateTag(`/posts/${params.id}`);
